Rename bower to to gallerygrid.js
Bower already contains a package called "gallerygrid". Since bower provides all kinds of stuff, not just js libraries, it makes sense to add the ".js" suffix to signal it being a js library. NPM does not need the suffix because it only manages js libraries. It's not the best solution to name the library differently in different package managers but I don't want to waste any more time to find a solution (e.g. a new name).